% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_test.R
\name{apply.test}
\alias{apply.test}
\title{Apply different tests to the simulated data}
\usage{
apply.test(sim.location, group, type='default',
subset=NA_real_, norm='pass',
test='wilcoxon', conf=NULL, ...)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{group}{character, name of the group within the .h5 file containing
the simulated features}

\item{type}{character, name of the index group within the .h5 file
containing the testing or bootstrapping indices
Can be either \code{"default"}, \code{"biomarker"},  \code{"confounder"},
or \code{"random"}}

\item{subset}{vector, subset size(s) to be tested}

\item{norm}{character, normalization method to be applied before testing}

\item{test}{character, differential abundance testing method}

\item{conf}{character, name of the confounder metadata column}

\item{...}{additional parameters (for experts)}
}
\value{
This function returns a list of P-values matrices
}
\description{
This function applies one of the implemented differential
abundance testing methods to a dataset within the simulation
}
\details{
todo
}
\keyword{SIMBA}
\keyword{apply.test}
