% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounder_plot.R
\name{confounder.plot}
\alias{confounder.plot}
\title{Plot ANOVA-type plots for original and simulated data}
\usage{
confounder.plot(sim.location, group)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{group}{character, name of the group within the .h5 file containing
the simulated features}
}
\value{
This function returns a list containing a ggplot object with a
}
\description{
This function will generate
}
\details{
This functions
TODO
}
\keyword{SIMBA}
\keyword{confounder.plot}
