% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biased_test_idx.R
\name{create.biased.idx}
\alias{create.biased.idx}
\title{Create biased indices for testing}
\usage{
create.biased.idx(sim.location, subsets, repetitions,
lab.svar = NULL, meta.svar = NULL, strat.scheme = 'all', bias = 0.5)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{subsets}{vector of intended subset sizes for resampled data}

\item{repetitions}{number of unique sets of indices to generate for each
subset size deemed valid}

\item{lab.svar}{for real data, name of the main stratification "label"
variable in the original metadata; defaults to \code{NULL} and is not
applicable for simulated data}

\item{meta.svar}{for real data, a single string or vector of strings
corresponding to the name(s) of binary metadata variable(s) by which biased
resampling should be stratified; defaults to \code{NULL} and is not
applicable for simulated data}

\item{strat.scheme}{character, name of the index group to be written within
the .h5 file containing different indices for downstream testing
Can be either \code{'biomarker'},\code{'confounder'}, \code{'random'},
or \code{'all'} (the default)}

\item{bias}{float, strength of the confounder bias in the subsampling}
}
\value{
vector of real metadata variables insufficient for resampling, if
applicable, otherwise nothing returned & indices are written to
\code{'sim.location'}
}
\description{
This function creates biased indices for testing with an
additional confounder
}
\details{
todo
}
\keyword{internal}
