% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_idx.R
\name{create.test.idx}
\alias{create.test.idx}
\title{Create indices for testing}
\usage{
create.test.idx(sim.location, subsets, repetitions)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{subsets}{vector of subset sizes to be sampled}

\item{repetitions}{integer, number of repetitions for each group/subset
combination}
}
\value{
Returns nothing, but modifies the simulation file
}
\description{
This function creates a matrix of test indices for each group
in the simulation file on which differential abundance testing methods can
be run
}
\keyword{SIMBA}
\keyword{create.test.idx}
