% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_test.R
\name{eval.test}
\alias{eval.test}
\title{Evaluate the results of different tests}
\usage{
eval.test(sim.location, group, res.mat, adjust='BH', alpha = 0.05)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{group}{character, name of the group within the .h5 file containing
the simulated features}

\item{res.mat}{matrix, output from the \link{apply.test} function}

\item{adjust}{character, indicate the multiple hypothesis testing to be
performed on the P-values, defaults to "BH"}

\item{alpha}{numeric, significance threshold at which the test will be
evaluated, defaults to \code{0.05}}
}
\value{
A dataframe containing the evaluation measures at the given alpha
value for the different test runs.
}
\description{
This function takes the results of the \link{apply.test}
function and calculates different evaluation measures (see Details).
}
\details{
This function will check that tests ran successfully and calculate
how well the supplied P-values can distinguish between the true differential
abudance features and background features. The measures computed are:
\itemize{
\item AUROC: measure for the separation of true and background features
\item TP: number of true positives detected at \code{alpha}
\item FP: number of false positives detected at \code{alpha}
\item TN: number of true negatives at \code{alpha}
\item FN: number of false negatives detected at \code{alpha}
\item PR: precision, calculated as TP/(FP+TP)
\item R: recall, calculated as TP/(TP+FN)
\item FDR: false discovery rate, calculated as FP/(FP+TP)
}
}
\keyword{SIMBA}
\keyword{eval.test}
