% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_sparseDOSSA.R
\name{fit.parameters}
\alias{fit.parameters}
\title{Fit the parameters from the original data}
\usage{
fit.parameters(feat.filt)
}
\description{
Fit the parameters from the original data
}
\keyword{internal}
