% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_sparseDOSSA.R
\name{funcMakeFeature}
\alias{funcMakeFeature}
\title{Generate feature matrix}
\usage{
funcMakeFeature(
  vdMu,
  vdSD,
  vdPercentZero,
  iNumberSamples,
  iMinNumberCounts,
  iMinNumberSamples,
  mdLogCorr = diag(length(vdSD)),
  vdTruncateThreshold = rep(Inf, length(vdSD)),
  fZeroInflate = TRUE,
  fVerbose = FALSE
)
}
\description{
Generate feature matrix
}
\keyword{internal}
