% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pcoa.R
\name{pcoa.plot}
\alias{pcoa.plot}
\title{Plot PCoAs for simulated data}
\usage{
pcoa.plot(sim.location, group)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{group}{character, name of the group within the .h5 file containing
the simulated features}

\item{distance}{character, name of the distance that should be used to
calculate the PCoA, should be either \code{"bray"} or \code{"log-euclidean"}}
}
\value{
This function returns a ggplot2 object with a PCoA plot, coloured by
group (either real or simulated data).
}
\description{
This function will generate a principal coordinate analysis
plot comparing the original data and the simulated data for different groups
}
\keyword{SIMBA}
\keyword{pcoa.plot}
