% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reality_check.R
\name{reality.check}
\alias{reality.check}
\title{Check simulations for realism}
\usage{
reality.check(sim.location, group, ml=FALSE)
}
\arguments{
\item{sim.location}{file name for the .h5 file containing the simulations}

\item{group}{character, name of the group within the .h5 file containing
the simulated features \strong{OR} \code{"original"} to get the same
measures for the original data set}

\item{ml}{Boolean, indicating if a machine-learning model should be trained
to distinguish real and simulated data (see Details below).
Defaults to \code{FALSE}}
}
\value{
Returns a list of measures
}
\description{
This function will compare simulated data with the
real data on which the simulations are based
}
\details{
This function computes a list containing different measures for how
realistic the simulations are, containing
\itemize{
\item \code{sparsity} measures for the sparsity of samples, computed as
L0-norm and as Gini-coefficient
\item \code{variance} measures for the variance of features, containing
additionally measures for differences between groups, such as generalized
fold change and prevalence difference
\item \code{AUC-BC} AUROC value for how well real and simulated data can be
distinguished based on the Bray-Curtis distance
\item \code{AUC-E} AUROC value for how well real and simulated data can be
distinguished based on the log-Euclidean distance
\item \code{AUC-ML} AUROC value for how well real and simulated data can be
distinguished based on a \link[SIAMCAT]{SIAMCAT}
machine learning model
}
}
\keyword{SIMBA}
\keyword{reality.check}
