% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_hawinkel.R
\name{rmvnegbin}
\alias{rmvnegbin}
\title{helper function to generate correlated counts}
\usage{
rmvnegbin(n, mu, Sigma, ks, ...)
}
\description{
helper function to generate correlated counts
}
\keyword{internal}
