% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_hawinkel.R
\name{simulate.betabin}
\alias{simulate.betabin}
\title{# wrapper for the beta-binomial simulations}
\usage{
\method{simulate}{betabin}(feat, meta, sim.out, sim.params)
}
\description{
# wrapper for the beta-binomial simulations
}
\keyword{internal}
