% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate.data}
\alias{simulate.data}
\title{Simulate metagenomic features}
\usage{
simulate.data(feat, meta, sim.location,
sim.method='resampling',
sim.type="cross-section",
filt.params=list(prev.cutoff=0.05, ab.cutoff=1e-04, log.n0=1e-05),
sim.params=list(class.balance=0.5, prop.markers=0.1,
ab.scale=c(1.0, 1.25, 1.5, 2.0, 5.0, 10.0, 20.0),
prev.scale=c(0.0, 0.1, 0.2, 0.3),
repeats=100, feature.type='all'))
}
\arguments{
\item{feat}{feature matrix, can be either a single matrix or a list of
matrices from different studies}

\item{meta}{metadata about samples in the features matrix, should be a
dataframe or a list of dataframes}

\item{sim.location}{filename for an .h5 file which will store the
simulations, original data, and simulation parameters}

\item{sim.method}{methodology to create the simulations. Can either be
a string describing a pre-implemented method or a user-specified function
(see Details below). Defaults to \code{"resampling"}}

\item{sim.type}{character, can be either \code{"cross-section"} or
\code{"time-course"}, defaults to \code{'cross-section'}}

\item{filt.params}{list of parameters for data filtering, defaults
to \code{filt.params=list(prev.cutoff=0.05, ab.cutoff=1e-04,
log.n0=1e-05)}, see Details below}

\item{sim.params}{list of parameters which are given to
the \code{sim.method} function, see Details}
}
\value{
Does not return anything but instead creates an h5 file
containing simulated samples
}
\description{
This function uses a feature and metadata table
as well as specific parameters to create simulated metagenomic samples
and stores those in the indicated .h5 file
}
\details{
This functions TODO

\bold{Filtering}
\bold{Pre-implemented Simulations}
\bold{Simulations}
\bold{h5-File Organisation}
}
\keyword{SIMBA}
\keyword{simulate.data}
