% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_hawinkel.R
\name{simulate.dirmult}
\alias{simulate.dirmult}
\title{# wrapper for the Dirichlet simulations}
\usage{
\method{simulate}{dirmult}(feat, meta, sim.out, sim.params)
}
\description{
# wrapper for the Dirichlet simulations
}
\keyword{internal}
