% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_hawinkel.R
\name{simulate.markers.dirmult}
\alias{simulate.markers.dirmult}
\title{# wrapper for the actual data generation}
\usage{
\method{simulate}{markers.dirmult}(pis, theta, libSizesOrig, label, markers, ab.scale)
}
\description{
# wrapper for the actual data generation
}
\keyword{internal}
