% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_hawinkel.R
\name{simulate.markers.negbin}
\alias{simulate.markers.negbin}
\title{# wrapper for the actual data generation}
\usage{
\method{simulate}{markers.negbin}(
  phis,
  rhos,
  libSizesOrig,
  label,
  markers,
  ab.scale,
  correlation = NULL
)
}
\description{
# wrapper for the actual data generation
}
\keyword{internal}
