% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_mmh.R
\name{simulate.MMH}
\alias{simulate.MMH}
\title{# wrapper for the McMurdie&Holmes simulations}
\usage{
\method{simulate}{MMH}(feat, meta, sim.out, sim.params)
}
\description{
# wrapper for the McMurdie&Holmes simulations
}
\keyword{internal}
