% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_hawinkel.R
\name{simulate.negbin}
\alias{simulate.negbin}
\title{# wrapper for the negative-binomial simulations}
\usage{
\method{simulate}{negbin}(feat, meta, sim.out, sim.params)
}
\description{
# wrapper for the negative-binomial simulations
}
\keyword{internal}
