% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_yang_chen.R
\name{simulate.SimMSeq}
\alias{simulate.SimMSeq}
\title{# wrapper for the Yang and Chen}
\usage{
\method{simulate}{SimMSeq}(feat, meta, sim.out, sim.params)
}
\description{
# wrapper for the Yang and Chen
}
\keyword{internal}
