% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_mmh.R
\name{simulate.W}
\alias{simulate.W}
\title{# wrapper for the Weiss simulations}
\usage{
\method{simulate}{W}(feat, meta, sim.out, sim.params)
}
\description{
# wrapper for the Weiss simulations
}
\keyword{internal}
