% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\encoding{UTF-8}
\name{toy.feat}
\alias{toy.feat}
\title{SIMBA example features}
\source{
\doi{10.1016/j.cell.2015.11.001}
}
\description{
Taxonomic profiles from the study by Zeevi et al. Cell 2014
(see \doi{10.1016/j.cell.2015.11.001}), 
containing 1952 features from 1181 samples, created by the motus profiler.

Mainly used for running the examples in the function documentation and
in the vignette.
}
\keyword{data}
