% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\encoding{UTF-8}
\name{toy.meta}
\alias{toy.meta}
\title{SIMBA example metadata}
\source{
\doi{10.1016/j.cell.2015.11.001}
}
\description{
Metadata from the study by Zeevi et al. Cell 2014
(see \doi{10.1016/j.cell.2015.11.001}).

Mainly used for running the examples in the function documentation and
in the vignette.
}
\keyword{data}
